Update Argon Rest Module
Adding new REST services
Resolving bugs for store data in BaseX
Adding class for versioning

+++ RestServicesModule/src/main/java/de/axxepta/resources/DocumentsResource.java
@@ -1,6 +1,5 @@
 package de.axxepta.resources;
 
-import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.IOException;
 import java.net.MalformedURLException;
@@ -31,14 +30,20 @@ import org.apache.log4j.Logger;
 import org.codehaus.plexus.util.FileUtils;
 
 import com.codahale.metrics.Meter;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
 
 import de.axxepta.dao.interfaces.IDocumentCacheDAO;
+import de.axxepta.dao.interfaces.IDocumentDAO;
 import de.axxepta.exceptions.ResponseException;
 import de.axxepta.models.FileDescriptionModel;
 import de.axxepta.models.FileDisplayModel;
 import de.axxepta.services.interfaces.IFileResourceService;
+import de.axxepta.tools.ExtractContentFile;
+import de.axxepta.tools.GetContentOfURL;
 import de.axxepta.tools.ValidationString;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 
@@ -65,7 +70,7 @@ public class DocumentsResource {
 	@ApiResponses({ @ApiResponse(responseCode = "200", description = "retrieve of json of uploaded files"),
 			@ApiResponse(responseCode = "409", description = "uploaded directory stored as constant not exist") })
 	@GET
-	@Path("list-uploaded-files")
+	@Path("list-dir-uploaded-files")
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response listUploadedFiles() throws ResponseException {
 		metricRegistry.mark();
@@ -77,6 +82,7 @@ public class DocumentsResource {
 		Map<String, FileDisplayModel> filesResponseMap = new HashMap<>();
 		for (File file : filesList) {
 			String key = FileUtils.basename(file.getName());
+			key = key.substring(0, key.length() - 1);
 			FileDisplayModel value = null;
 			try {
 				value = new FileDisplayModel(file);
@@ -93,31 +99,33 @@ public class DocumentsResource {
 	@ApiResponses({ @ApiResponse(responseCode = "200", description = "change database"),
 			@ApiResponse(responseCode = "406", description = "request is do it from address other than local address"),
 			@ApiResponse(responseCode = "409", description = "not exist database with name wanted") })
-	
+
 	@PUT
 	@Path("set-databaseName")
-	public Response setDatabaseName(@QueryParam("name") String databaseName) {
+	public Response setDatabaseName(
+			@Parameter(description = "name of the database", required = true) @QueryParam("database-name") String databaseName) {
 		metricRegistry.mark();
-		
-		if(!servletRequest.getRemoteAddr().equals("127.0.0.1")) {
+
+		if (!servletRequest.getRemoteAddr().equals("127.0.0.1")) {
 			LOG.error("The service has been attempted to be runned by " + servletRequest.getRemoteAddr());
 			return Response.status(Status.NOT_ACCEPTABLE).entity("That service must be used only locally").build();
 		}
-		
+
 		LOG.info("Trying to set new database with name " + databaseName);
-		
+
 		Boolean r = documentCacheDAO.setDatabaseName(databaseName);
-		
-		if(r == null) {
-			return Response.status(Status.ACCEPTED).entity("Database with name " + databaseName + " is already setted").build();
+
+		if (r == null) {
+			return Response.status(Status.ACCEPTED).entity("Database with name " + databaseName + " is already setted")
+					.build();
 		}
-		
-		if(!r) {
-			return Response.status(Status.CONFLICT).entity("Database with name "  + databaseName + " not exists").build();
-		}
-		else {
+
+		if (!r) {
+			return Response.status(Status.CONFLICT).entity("Database with name " + databaseName + " not exists")
+					.build();
+		} else {
 			LOG.info("Database name is changed in " + databaseName);
-			return Response.status(Status.OK).entity("New database used have name "  + databaseName).build();
+			return Response.status(Status.OK).entity("New database used have name " + databaseName).build();
 		}
 	}
 
@@ -127,7 +135,10 @@ public class DocumentsResource {
 			@ApiResponse(responseCode = "400", description = "name of file cannot be valided") })
 	@GET
 	@Path("exist-file")
-	public Response existFile(@QueryParam("filename") String fileName) throws ResponseException {
+	public Response existFile(
+			@Parameter(description = "file name", required = true) @QueryParam("filename") String fileName,
+			@Parameter(description = "is stored locally") @QueryParam("stored-as-temp") boolean storedAsTemp)
+			throws ResponseException {
 		LOG.info("test if file service");
 		metricRegistry.mark();
 		if (!ValidationString.validationString(fileName, "fileName")) {
@@ -142,17 +153,20 @@ public class DocumentsResource {
 			return Response.status(Status.ACCEPTED).entity("File  with name " + fileName + " not exist").build();
 	}
 
-	@Operation(summary = "Upload file", description = "Upload file as an temporal one", method = "POST", operationId = "#4_4")
+	@Operation(summary = "Upload file", description = "Upload file in directory and maybe also in database", method = "POST", operationId = "#4_4")
 	@ApiResponses({
 			@ApiResponse(responseCode = "200", description = "upload file succes and return URL for temporal file and his type in JSON"),
 			@ApiResponse(responseCode = "400", description = "at least one parameter in the request is missing"),
 			@ApiResponse(responseCode = "409", description = "temporal file cannot be created") })
 	@POST
-	@Path("upload")
+	@Path("upload-file")
 	@Consumes(MediaType.APPLICATION_FORM_URLENCODED)
 	@Produces(MediaType.APPLICATION_JSON)
-	public Response uploadFileLocal(@FormParam("file-url") String fileURLString,
-			@DefaultValue("true") @FormParam("as-temp-file") boolean isTmpFile) throws ResponseException {
+	public Response uploadFile(
+			@Parameter(description = "URL of the file", required = true) @FormParam("file-url") String fileURLString,
+			@Parameter(description = "is stored in the database") @FormParam("is-added-to-database") boolean isAddedToDatabase,
+			@Parameter(description = "is stores as a temp file") @FormParam("is-temp-file") boolean isTmpFile)
+			throws ResponseException {
 		metricRegistry.mark();
 
 		if (!ValidationString.validationString(fileURLString, "file url")) {
@@ -169,21 +183,54 @@ public class DocumentsResource {
 			throw new ResponseException(Response.Status.BAD_REQUEST.getStatusCode(),
 					"String does not determine a valid URL");
 		}
+
 		FileDescriptionModel fileDescription = fileService.uploadLocalFile(fileURL, isTmpFile);
 
 		if (fileDescription == null)
-			throw new ResponseException(Response.Status.CONFLICT.getStatusCode(), "The file could not be uploaded");
+			throw new ResponseException(Response.Status.CONFLICT.getStatusCode(),
+					"An error occour in uploaded file in directory");
+
+		if (isAddedToDatabase) {
+
+			String content;
+			try {
+				content = GetContentOfURL.getContent.getContent(new URL(fileURLString));
+			} catch (MalformedURLException e) {
+				LOG.error(fileURLString + " malformed URL");
+				throw new ResponseException(Response.Status.CONFLICT.getStatusCode(), "Malformed URL");
+			}
+
+			String fileName = fileURLString.substring(fileURLString.lastIndexOf("/") + 1);
+
+			if (content == null) {
+				LOG.error("Exception during getting content of page");
+				throw new ResponseException(Response.Status.CONFLICT.getStatusCode(),
+						"Exception during getting content of page");
+			}
+
+			boolean result = documentCacheDAO.save(fileName, content);
+
+			if (result) {
+				fileDescription.setAddedToDatabase(true);
+			}
+		} else {
+			LOG.info("File from URL " + fileURLString + " will not be added to the database");
+		}
+
 		return Response.ok(fileDescription).build();
 	}
 
-	@Operation(summary = "Delete a file", description = "Delete an file from uploaded set", method = "DELETE", operationId = "#4_5")
+	@Operation(summary = "Delete a file", description = "Delete an file from upload directory and from database", method = "DELETE", operationId = "#4_5")
 	@ApiResponses({ @ApiResponse(responseCode = "200", description = "the file was successfully deleted"),
 			@ApiResponse(responseCode = "202", description = "file not exist"),
 			@ApiResponse(responseCode = "409", description = "name of file cannot be validated") })
 	@DELETE
 	@Path("delete-file")
 	@Produces(MediaType.TEXT_PLAIN)
-	public Response deleteFile(String fileName) throws ResponseException {
+	public Response deleteFile(
+			@Parameter(description = "file name", required = true) @QueryParam("filename") String fileName,
+			@Parameter(description = "is stored in the database") @QueryParam("is-from-database") @DefaultValue("false") boolean isFromDatabase)
+			throws ResponseException {
 		LOG.info("delete file service");
 		metricRegistry.mark();
 
@@ -191,22 +238,40 @@ public class DocumentsResource {
 			LOG.error("Not valid file name");
 			throw new ResponseException(Response.Status.BAD_REQUEST.getStatusCode(), "Not valid file name");
 		}
-		boolean hasDeleted = fileService.deleteFile(fileName);
+
+		String response = "";
+		boolean hasDeleted = false;
+
+		if (!isFromDatabase) {
+			LOG.info(fileName + " will not be deleted from database");
+			hasDeleted = fileService.deleteFile(fileName);
+
+			if (hasDeleted) {
+				response = "File  with name " + fileName + " was deleted from upload directory";
+			} else {
+				response = "File with name " + fileName + " wasn't deleted from upload directory";
+			}
+		} else {
+			LOG.info(fileName + " will be deleted from database");
+			response = "file will be deteleted from database";
+		}
+
 		if (hasDeleted)
-			return Response.status(Status.OK).entity("File  with name " + fileName + " was deleted").build();
+			return Response.status(Status.OK).entity(response).build();
 		else
-			return Response.status(Status.ACCEPTED).entity("File with name " + fileName + "has not been deleted")
-					.build();
+			return Response.status(Status.ACCEPTED).entity(response).build();
 	}
 
-	@Operation(summary = "Retrieve file", description = "Retrieve file as an binary one", method = "GET", operationId = "#4_6")
+	@Operation(summary = "Retrieve file", description = "Retrieve file as an binary one from upload directory or from database", method = "GET", operationId = "#4_6")
 	@ApiResponses({ @ApiResponse(responseCode = "200", description = "return file content"),
 			@ApiResponse(responseCode = "409", description = "name of file cannot be validated") })
 	@GET
 	@Path("retrieve-file")
-	@Produces(MediaType.APPLICATION_OCTET_STREAM)
-	public Response retrieveFile(@QueryParam("filename") String fileName) throws ResponseException {
-		LOG.info("retrieve file service");
+	public Response retrieveFile(
+			@Parameter(description = "file name", required = true) @QueryParam("filename") String fileName,
+			@Parameter(description = "is from database") @QueryParam("is-from-database") boolean isFromDatabase)
+			throws ResponseException {
+		LOG.info("retrieve temp file service");
 		metricRegistry.mark();
 
 		if (!ValidationString.validationString(fileName, "fileName")) {
@@ -214,11 +279,49 @@ public class DocumentsResource {
 			throw new ResponseException(Response.Status.BAD_REQUEST.getStatusCode(), "File name is not valid");
 		}
 
-		byte[] bytesFile = fileService.readingFile(fileName);
-		ByteArrayOutputStream boStream = new ByteArrayOutputStream(bytesFile.length);
-		boStream.write(bytesFile, 0, bytesFile.length);
-		return Response.ok(boStream, "image/png").header("content-disposition", "attachment; filename = " + fileName)
-				.build();
+		long startCall = System.nanoTime();
+		String content = documentCacheDAO.getContentFile(fileName);
+		long endCall = System.nanoTime();
+
+		LOG.info("Time(in nanoseconds) for retrieving file " + (endCall - startCall) + " with bytes buffer size ");
+
+		LOG.info("Value content " + content);
+		
+		if (content == null) {		
+			return Response.ok("File " + fileName + " not found").build();
+		}
+		
+		byte[] bytesFile = content.getBytes();
+
+		return Response.ok(bytesFile, MediaType.APPLICATION_OCTET_STREAM)
+				.header("content-disposition", "attachment; filename = " + fileName).build();
+
+	}
+	
+	@Operation(summary = "Retrieve list of files from one of database", description = "Retrieve list of files stored in setted BaseX database", method = "GET", operationId = "#4_7")
+	@ApiResponses({ @ApiResponse(responseCode = "200", description = "return list of file names"),
+		@ApiResponse(responseCode = "409", description = "list of file names cannot be obtain") })
+	@GET
+	@Path("list-file-names-from-database")
+	public Response getListNameFileDatabase(@Parameter(description = "is from cache database") @QueryParam("is-from-cache") boolean isFromCacheDatabase) throws ResponseException {
+		metricRegistry.mark();
+		
+		List<String> fileNameList = documentCacheDAO.getListFileName(isFromCacheDatabase);
+		
+		if(fileNameList == null) {
+			LOG.error("list of files cannot be obtain from database");
+			throw new ResponseException(Response.Status.CONFLICT.getStatusCode(), "list of files cannot be obtain from database");
+		}
+		
+		ObjectMapper mapper = new ObjectMapper();
+		String jsonFileNames = null;
+		try {
+			jsonFileNames = mapper.writeValueAsString(fileNameList);
+		} catch (JsonProcessingException e) {
+			
+		}
+		
+		return Response.status(Status.OK).entity(jsonFileNames).build();
 	}
 	
 }
